whitespace cleanup
Trying to make the use of whitespace more consistent and the files
more readable (it's a mess if tabs and spaces are mixed).

As the prevalent intendation style was using spaces, all files are now
using spaces for indentation. Within each file the indentation size is
consistent, for some files it's 4, for others 2 spaces.

Below you find a summary and the used a python regexp usful for quick spotting:
- trailing whitespace: (?<!\s)[\t ]+$
- tabs or multi whitespace in lines with other chars: (?<=\S)([\t ]*\t[\t ]*|[\t ]* {2,}[\t ]*)(?=\S)
- mixed space / tab indents unified per file: ^( +\t+\s+)